apps/api/pyproject.toml
[tool.poetry]
name = "taskflow-api"
version = "0.1.0"
description = "TaskFlow FastAPI backend"
authors = ["Guy Romelle Magayano <aspiredtechie2010@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
strawberry-graphql = { extras = ["fastapi"], version = "^0.223.0" }
sqlalchemy = "^2.0.35"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["argon2"], version = "^1.7.4" }
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
redis = { extras = ["hiredis"], version = "^5.2.0" }
structlog = "^24.4.0"
prometheus-client = "^0.21.0"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
testcontainers = { extras = ["postgresql"], version = "^4.6.0" }
factory-boy = "^3.3.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
types-redis = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=html --cov-report=term"
